% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_predictor.R
\name{MP_predictor}
\alias{MP_predictor}
\title{MP predictor function}
\usage{
MP_predictor(
  drug_names,
  ID_url = "https://www.ebi.ac.uk/chembl/api/data/molecule/search",
  base_url = "https://www.ebi.ac.uk/chembl/api/data/activity",
  limit = 10000
)
}
\arguments{
\item{drug_names}{The name of the drug of interest or input custom to build the dataset from scratch}

\item{ID_url}{= "https://www.ebi.ac.uk/chembl/api/data/molecule/search" Non-editable URL, required to pull CHEMBL IDs}

\item{base_url}{= "https://www.ebi.ac.uk/chembl/api/data/activity" Non-editable URL, required to pull unbound fractions and activity data}

\item{limit}{= 10000 Required to scan through the CHEMBL API pages}
}
\value{
A data frame with MP predictions from the phase distribution equation, Koshimichi et al. equation, logarithmically transformed PD equation, Meskin and Lien equation, and
XGBoost model
}
\description{
The function calculates the milk-to-plasma ratio (MP) using information pulled from the CHEMBL database API. It will acquire
the data required to predict the phase distribution equation, the equation proposed by Koshimichi et al., the logarithmically transformed
phase distribution equation, Meskin and Lien's equation, and finally the model trained using the XGBoost ML method. The relevant equations
can be found online or by viewing the equations nested within the function. Drug names or "custom" are all that is required to use the function
and the user is able to modify any of the input parameters as they see fit. The goal of this function is to facilitate MP predictions
and to provide the user with the choice of MP to use in PBPK models.
}
\details{
Predict the MP using information from CHEMBL

Enter the name of the drug of interest. The first step is to go through CHEMBL
and pull data from databases called resources. An entry may or may not have a record in a
particular resource.


If you are unsure about which name to use, you could go onto the CHEMBL website: https://www.ebi.ac.uk/chembl/
and find the drug of interest. Then enter the name into the function and the function should work as intended.
}
\note{
Links to the webservice documentation:
\itemize{
     \item \url{https://www.ebi.ac.uk/chembl/api/data/docs}
}
}
